xml/parsing/manakai/entrefs-5.dat

#data
<!DOCTYPE a[
  <!ENTITY b PUBLIC "aa" "bb">
]><a>&b;</a>
#errors
3;6;external entref;;b;;i
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
  <!ENTITY b PUBLIC "aa" "bb">
  <!ENTITY c "e&b;d">
]><a>&c;</a>
#errors
1;2;external entref;;b;;i
#document
| <!DOCTYPE a>
| <{} a>
|   "ed"

#data
<!DOCTYPE a[
  <!ENTITY b PUBLIC "aa" "bb" NDATA x>
]><a>&b;</a>
#errors
3;6;unparsed entity;;b;;m
#document
| <!DOCTYPE a>
|   <!ENTITY b "" "aa" "bb" x>
| <{} a>
|   "&b;"
